# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BrandResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      logo_url:
        type: string
      name:
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
      workflow_state:
        $ref: '#/definitions/dto.StateView'
    type: object
...
  dto.CategoryResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      level:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      slug:
        type: string
    type: object
...
  dto.CreateProductAnswerRequest:
    properties:
      body:
        maxLength: 2000
        minLength: 2
        type: string
    required:
    - body
    type: object
  dto.CreateProductQuestionRequest:
    properties:
      body:
        maxLength: 1000
        minLength: 5
        type: string
    required:
    - body
    type: object
...
  dto.PriceHistoryPoint:
    properties:
      compare_at_price:
        type: number
      price:
        type: number
      recorded_at:
        type: string
    type: object
  dto.ProductAlternativeResponse:
    properties:
      allow_backorder:
        type: boolean
      attributes:
        items:
          $ref: '#/definitions/dto.ProductAttributeResponse'
        type: array
      barcode:
        type: string
      brand:
        $ref: '#/definitions/dto.BrandResponse'
      brand_id:
        type: integer
      category:
        $ref: '#/definitions/dto.CategoryResponse'
      category_id:
        type: integer
      channels:
        items:
          type: string
        type: array
      colors:
        items:
          type: string
        type: array
      compare_at_price:
        type: number
      cost:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_digital:
        type: boolean
      is_new:
        type: boolean
      low_stock_threshold:
        type: integer
      meta_description:
        type: string
      meta_title:
        type: string
      name:
        type: string
      price:
        type: number
      published_at:
        type: string
      rating:
        type: number
      reviews_count:
        type: integer
      sizes:
        items:
          type: string
        type: array
      sku:
        type: string
      slug:
        type: string
      status:
        type: string
      stock:
        type: integer
      store:
        $ref: '#/definitions/dto.ProductStoreSummary'
      store_id:
        type: integer
      store_logo:
        type: string
      store_name:
        type: string
      store_rating:
        type: number
      store_slug:
        type: string
      tags:
        items:
          type: string
        type: array
      track_inventory:
        type: boolean
      updated_at:
        type: string
      visibility:
        description: Publishing extras
        type: string
      warehouse_location:
        type: string
      weight:
        type: number
      workflow_state:
        $ref: '#/definitions/dto.StateView'
    type: object
  dto.ProductAnswerResponse:
    properties:
      author:
        type: string
      body:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_ai_reply:
        type: boolean
      is_store_reply:
        type: boolean
      question_id:
        type: integer
    type: object
...
  dto.ProductAttributeResponse:
    properties:
      name:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
  dto.ProductQuestionResponse:
    properties:
      answers:
        items:
          $ref: '#/definitions/dto.ProductAnswerResponse'
        type: array
      author:
        type: string
      body:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_owner:
        type: boolean
      product_id:
        type: integer
    type: object
...
  dto.ProductStoreSummary:
    properties:
      follower_count:
        type: integer
      id:
        type: integer
      is_verified:
        type: boolean
      location:
        type: string
      logo_url:
        type: string
      name:
        type: string
      rating:
        type: number
      return_policy:
        type: string
      review_count:
        type: integer
      shipping_info:
        type: string
      slug:
        type: string
    type: object
...
  dto.StateView:
    properties:
      code:
        type: string
      color:
        type: string
      id:
        type: integer
      is_final:
        type: boolean
      is_initial:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      text_color:
        type: string
    type: object
  dto.StockNotificationStatusResponse:
    properties:
      subscribed:
        type: boolean
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@luxe.com
    name: API Support
  description: Production-grade e-commerce backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Shopping Platform API
  version: "1.0"
paths:
...
  /products/{id}/alternatives:
    get:
      description: Lists the same product model from other stores (matched by barcode)
      parameters:
      - description: Product ID or slug
        in: path
        name: id
        required: true
        type: string
      - default: 6
        description: Max items
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    alternatives:
                      items:
                        $ref: '#/definitions/dto.ProductAlternativeResponse'
                      type: array
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get cross-store product alternatives
      tags:
      - Products
...
  /products/{id}/price-history:
    get:
      description: Returns price snapshots over time for PDP charts
      parameters:
      - description: Product ID or slug
        in: path
        name: id
        required: true
        type: string
      - default: 90
        description: Number of days to include
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    points:
                      items:
                        $ref: '#/definitions/dto.PriceHistoryPoint'
                      type: array
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get product price history
      tags:
      - Products
  /products/{id}/questions:
    get:
      description: Paginated buyer questions and answers for a product
      parameters:
      - description: Product ID or slug
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    limit:
                      type: integer
                    offset:
                      type: integer
                    questions:
                      items:
                        $ref: '#/definitions/dto.ProductQuestionResponse'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List product Q&A
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Post a buyer question on the product detail page
      parameters:
      - description: Product ID or slug
        in: path
        name: id
        required: true
        type: string
      - description: Question body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductQuestionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductQuestionResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Ask a product question
      tags:
      - Products
  /products/{id}/questions/{questionId}/answers:
    post:
      consumes:
      - application/json
      description: Post an answer; store owners are marked as store replies
      parameters:
      - description: Product ID or slug
        in: path
        name: id
        required: true
        type: string
      - description: Question ID
        in: path
        name: questionId
        required: true
        type: integer
      - description: Answer body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductAnswerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductAnswerResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Answer a product question
      tags:
      - Products
...
  /products/{id}/stock-notifications:
    delete:
      parameters:
      - description: Product ID or slug
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StockNotificationStatusResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Unsubscribe from back-in-stock notifications
      tags:
      - Products
    get:
      parameters:
      - description: Product ID or slug
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StockNotificationStatusResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get back-in-stock subscription status
      tags:
      - Products
    post:
      parameters:
      - description: Product ID or slug
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StockNotificationStatusResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Subscribe to back-in-stock notifications
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"